Update to Index
I updated the index with Data Sprint Goals for sites which are new.  The idea is that they would read this as a starting point and then be led to a Sprint Page in the GitHub.  I also added an EHR Uploading header because I felt there should be something here for that also?  Not sure what would be a good fit but I didn't want to the front page to be empty.