[DC-278] Add script for data retraction (#119)
Add `retract_data.py`, a script which, for data retraction given a `person_id` and bucket name, removes all rows referencing the `person_id` in all CDM and PII files. It will do this for all folders in the bucket by default or one can optionally supply a folder to do this on. The main function returns a dictionary whose keys are folder names and values are lists of metadata asociated with updated GCS files.