[DC-336]  Load rdr etl into bigquery
While loading the rdr data into bigquery, the observation_period
csv broke the script.  This updates the script to allow jagged
line csv with trailing nullable fields for the observagtion_period
csv.  It moves the nullable timestamp field definitions to the end
of the json definition file and updates the command line as needed.