[DC-355] fix errors related to top_heel_errors report
This patch applies pylint changes to the python file.
Creates a dev_requirements file to load pylint and
coverage.  They are not need in production, so do
not belong in the production requirements file.  The name
in the bash script usage line is updated to match the name
of the script.  Once environment variables are set, the
option to set the variables is turned off.  Script indicates
if the file is created.