[DC-343] Pdf dump validation (#135)
* DC-343 Fix validation pipeline to handle PDF dump file

This patch creates a directory ignore list.  This list is
used to determine a submission directory during validation.
If a directory is skipped, the skip is logged.

- Test adding PDF files to site buckets

This patch cleans out some debug print messages, fixes a bug
in the submission directory constant for 'participant', improves
the comment for _get_submission_folder.

- Update pipeline to handle PDF file dump

This patch lets the internal copy_files function
ignore the Participant/ directory using a case insensitive
match.  It requires the ignored directory to be named
Participant and to exist in the root directory of the
bucket.