DC-356  develop id validation report
Before the validation report can be generated, the data has to
be read into the module.  The update to bq_utils allows paging
through a large result set.  The search and validation loop is
changed from a site based loop to an information object based
loop to allow garbage collection to clean memory as needed.  It
will no longer hold all the results for all the data in memory
for the duration of the validation.  This is an extensive refactor.
Constants should be used and should reside in their own modules.
They have been pulled into their own package.