DC-356  Develop id match report
This breaks the overly long identity_match module into four separate
modules for reading, writing, normalizing, and comparing identity values.
It updates the match tables to contain a value for an algorithm comparison.
It generates a report for a list of hpos in the internal bucket.  Still
needs to generate reports for individual hpos and implement unit tests.