DC-363  Failing unit tests for html report
The unit tests are failing due to the current directory structure.
Python is looking for the tools package in the current directory first, finding
it, and looking there for modules that are actually defined in data_steward/tools.
This adds the root directory to the unit_test/ package to eliminate this
ambiguity.