[DC-363] get consolidated html results from sites (#151)
* * Updates Consolidated achilles script to get the latest reports  …
* Adds script to get latest results.html from the site buckets

* * Updates Consolidated achilles script to get the latest reports

* DC-363  Adding some unit tests for consolidated html results script

This adds some unit tests for one module in the consolidated_reports
package.  It is not exhaustive.  It is a starting point for unit
testing the scripts in the consolidated_reports package.

* DC-363  Fixing codacy errors in unit test module

Removed unused imports and trailing whitespace.

* DC-363  Failing unit tests for html report

The unit tests are failing due to the current directory structure.
Python is looking for the tools package in the current directory first, finding
it, and looking there for modules that are actually defined in data_steward/tools.
This adds the root directory to the unit_test/ package to eliminate this
ambiguity.