DC-407  develoment enhancements
This adds the ability to log to the console or to a file in /tmp.
Adds the ability to parse the command line if run as an independent module.
Adds structure to name the queries and store a list of query names in the
constants files.  This will make modifications easier in the future.