[DC-417] Deid 02 (#220)
* Deid-Engine 2.0

DC-167 
DC-292 
DC-264
DC-417
DC-247
DC-248
 save string types with NULL as the value
Filter on the observation_concept_id and add in a missing observation_source_value.
Support rules with identifiers, and added rules with identifiers adding rules for both ids and values
This allows deid generalizations to work for the id based configuration files.
adds 'like' and 'not like' row suppression abilities
NOT IN cannot be used with nullable columns.
The config files were adjusted and the press and rules files
were adjusted to allow using exists/not exists in the config
files to search/update on nullable columns.
This updates the batch scripting of deid and
uses bq_utils to generate the output tables.  This
ensures the tables are defined as expected.
Removing old modules
Removes dead code
Adds library to requirements file
This will allow a person to write 'f a s t tests' (without the extra spaces)
in a commit message.
Only the fast unit tests will execute.  Should not be used as part
of a PR.