[DC-292] De id copy generalizations (#251)
* DC-292 Generalization is not occuring across multiple fields

This QA check failed because one field was generalized while
the other was not.  As of now, we want the two fields to match.
This adds the ability to set the value of field_1 based
on the new, generalized value of field_2.  It also throws in
a bit of non-deterministic shuffling of research_ids and date shifts.

* Breaking the lists into reasonable line segments

A 10,000 character line is too long.  This breaks
the json lists into reasonable line lengths.