[DC-544]/[DC-512] Generalizing Participant PIIState when EHR records are from another state
Uses standard python logging in favor of the current logging defined for the AOU class only.
Moves supporting table creation out of the intialization module to support maintainability.
Allows skipping supporting table creation if not needed to be applied to a table (tables
for observation de-identification only need to be created when de-identifying the observation table).
Creates a working case statement for state generalizations.
The config file creates a rule that can be successfully
applied to de-identify the PIIState for participants whose
EHR source state does not match their PIIState.