Add the ability to run DML statements on de-identified tables
* Builds in the ability to run a DML statement to remove duplicate
records after performing a generalization on data with multiple
answer choices.
* Refactors how default rules are processed, making it easier to identify which rules
are manipulated and how.
* Allows easier where clause conditions and automatically incorporates fixes for
conditions on nullable columns.  This should allow the clause to be read
so that a duplicate drop clause can be created from the existing information.
* Adding query configuration dictionary to avoid possible SQL injection messages thrown by Codacy.
* Using bq_utils to get fields, instead of a query.
* This should allow multiple dml statements to be created and executed.
There will likely need to be some code around the rules as they run,
but the configuration is in place.
* Adds a single unit test as a beginning point for unit testing deid.