[EDQ-417][EDQ-419] 'Visit Date Disparity' notebooks and changes (#462)
- Changed scripts so the dataset is imported from the notebooks module - never explicitly written in.
- Finished with the procedure, observation, measurement, drug, and condition table queries and calculations.   
   - Successfully merged all of the dataframes and outputted to a CSV   
- Created a valid 'other scripts' version of the date disparity check. Essentially removed the unioned_ehr prefixes.
- Made appropriate changes to m_o_t notebook and verified on all 6 fronts.

Need to switch to another branch. Shallow commit as of right now.

Improved docstrings. Changed .gitignore to prevent files used for dataframe generation from making their way into any commits.

Changed .gitignore and appropriately staged.

Changed the dq_issue_tracker to by a Python package so the 'main script' can leverage a variety of files and import functions.

Streamlined files in the greater Python package.

Added HPO class defintion and a function to identify the problem issues.

Added several new functions and continually adding to the dictionaries used for the indexing.

Renamed the main to create_hpo_objects to allow for easier processing down the line.

Still in progress

Verified that the right error messages are populating for CUMC.

Changed file names accordingly.

Improvements to the docstrings and naming conventions.

Finished all of the docstrings for the functions.

Marginal improvements in docstrings.

First iteration that outputs a set of files. Still no means to check when the issue was first instantiated. Still not validated.

Verified for Columbia and Banner. Still need to implement a cross-reference for 'first-reported.

Started to work on 'cross_reference_old_metrics. More work than anticipated but still coming along.

Started first implementation of 'cross reference functions to attempt to replace the 'first reported' metric dilemma. Have not troubleshot yet. Also changed the gitignore to exclude text files that are useful for documenting testing notes.