[EDQ-462] Small Fixes in Data Quality Metrics Processes
temporary fix for duplicates error. resolving the nan issue in the aggregate objects.


Changed the default of 'nan' to '0' for the data quality metrics (when instantiating). Found in the get_info function.

fixed the dictionaries and lists with respect to the constants file.

determined the types of metrics that neeed to be flipped. implemented the logic changes in terms of subtracting from 100 and changing the 'target low'

Inverted the percentages for the appropriate data quality metrics. Still need to figure out how to get the full site name.

changed the default value from '100' to '0'

starting to fix the combined_deid scripts for the site_df names

Changed the means of generating the full site names in the combined/deid datasets.

running a retrospective on the datasets for upload to the google drive.

fixed the issue of datetime casting in one of the notebooks.