[DC-1040] cope survey versioning
This is left as a manual cleaning rule script, but it's dependence
on the clean engine has been intentionally broken.  The clean engine is being udpated
and this would add extra complexity to that udpate.  Since this script will likely
be refactored later, the clean_engine components were commented out
and the script was set to manually execute the cleaning rules.