[DC-1394] Too much survey data being suppressed/cleaned
This changes the run order of several cleaning rules in the RDR data
stage.  This will fix the issue that caused the smoking data to be
suppressed.

It adds comments to help identify some common expected rule running orders.

It removes some useless code from the manual cleaning rules that prevented
the rules from being run as standalone rules.  They ran fine as part of the
RDR list.